+++ zcash_client_backend/src/tor/grpc.rs
@@ -1,4 +1,5 @@
 use std::{
+    error::Error as _,
     fmt,
     future::Future,
     pin::Pin,
@@ -20,6 +21,8 @@ impl Client {
         &self,
         endpoint: Uri,
     ) -> Result<CompactTxStreamerClient<Channel>, Error> {
+        self.ensure_bootstrapped().await?;
+
         let is_https = http::url_is_https(&endpoint)?;
 
         let channel = Endpoint::from(endpoint);
@@ -86,7 +89,15 @@ pub enum GrpcError {
 impl fmt::Display for GrpcError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
-            GrpcError::Tonic(e) => write!(f, "Hyper error: {}", e),
+            GrpcError::Tonic(e) => {
+                if let Some(source) = e.source() {
+                    // Tonic doesn't include the source error in its `Display` impl;
+                    // add it manually for the benefit of our downstreams.
+                    write!(f, "Tonic error: {e}: {source}")
+                } else {
+                    write!(f, "Tonic error: {e}")
+                }
+            }
         }
     }
 }