+++ zcash_client_backend/src/tor/http/cryptex.rs
@@ -14,7 +14,12 @@ mod gemini;
 mod ku_coin;
 mod mexc;
 
+/// Maximum number of retries for exchange queries implemented in this crate.
+const RETRY_LIMIT: u8 = 1;
+
 /// Exchanges for which we know how to query data over Tor.
+///
+/// Queries to these exchanges will be retried a single time on error.
 pub mod exchanges {
     pub use super::binance::Binance;
     pub use super::coinbase::Coinbase;
@@ -135,6 +140,8 @@ impl Client {
         &self,
         exchanges: &Exchanges,
     ) -> Result<Decimal, Error> {
+        self.ensure_bootstrapped().await?;
+
         // Fetch the data in parallel.
         let res = join!(
             exchanges.trusted.query_zec_to_usd(self),